Make cinder volume optional
In the swarm_atomic and k8s_atomic drivers container images are
stored in a dedicated cinder volume per cluster node. It is
proven that this architecture can be a scalability bottleneck.

Make the use of cinder volumes for container images and opt-in
option.  If docker-volume-size is not specified no cinder
volumes will be created. Before, if docker-volume-size wasn't
specified the default value was 25.

To use cinder volumes for container storage the user will
interact with magnum as before, (meaning the valid values are
integers starting from 1).

Closes-Bug: #1638006

+++ magnum/drivers/common/templates/fragments/configure_docker_storage_driver_atomic.sh
@@ -15,9 +15,11 @@ configure_overlay () {
 
     rm -rf /var/lib/docker/*
 
-    mkfs.xfs -f ${device_path}
-    echo "${device_path} /var/lib/docker xfs defaults 0 0" >> /etc/fstab
-    mount -a
+    if [ -n "$DOCKER_VOLUME_SIZE" ] && [ "$DOCKER_VOLUME_SIZE" -gt 0 ]; then
+        mkfs.xfs -f ${device_path}
+        echo "${device_path} /var/lib/docker xfs defaults 0 0" >> /etc/fstab
+        mount -a
+    fi
 
     echo "STORAGE_DRIVER=overlay" > /etc/sysconfig/docker-storage-setup
 
@@ -31,8 +33,10 @@ configure_overlay () {
 configure_devicemapper () {
     clear_docker_storage_congiguration
 
-    pvcreate -f ${device_path}
-    vgcreate docker ${device_path}
+    if [ -n "$DOCKER_VOLUME_SIZE" ] && [ "$DOCKER_VOLUME_SIZE" -gt 0 ]; then
+        pvcreate -f ${device_path}
+        vgcreate docker ${device_path}
 
-    echo "VG=docker" > /etc/sysconfig/docker-storage-setup
+        echo "VG=docker" > /etc/sysconfig/docker-storage-setup
+    fi
 }
